gitr: Update default git version to be cloned
A few of the git regression tests that passed on Fedora 39 hosts now
fail on Fedora 40. This seems to have been addressed in git-2.46.0,
the latest stable git version as of this writing.

But the newer unit tests seem to require more disk space than the
meager 5GB currently used as the default.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>