playbooks: Add a Code Ready Builder role
Installing Code Ready Builder turns out to be complicated by the way
the package is named on various cloud providers. Since there is more
than one workflow that wants to install CRB, move the installation
mechanism to its own role so that can be included in each workflow
that might need it.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>