nfsd: Detect the /data partition directly
The nfsd playbook runs its set-up during "bringup", but the /data
partition is set up later when the kernel is built and workflows are
set up. This means the nfsd playbook has to avoid using one of the
block devices so that later steps can put the /data partition on it.

Instead of counting the number of available block devices and just
skipping one of them, explicitly avoid using the device listed in
{{ data_device }}. This is a more reliable check.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>