terraform/Azure: Enable the use of client secrets
Add an alternate authentication mechanism for Azure: service
principal with client secret. The authentication identity and
secrets are maintained in shell environment variables.

Clean up: Move the authentication choices to the top of the
Kconfig submenu, since those have to be working first before
any of the other menu selections become meaningful.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>