terraform/Azure: Expand the list of locations
Add common US region codes that are useful to kdevops users not
located on the west coast. This patch also prepares for adding
region codes for other continents.

Moving region selection to a separate file could make it easier to
eventually generate this menu by running a script to query Azure
and build the menu entries.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>