create_data_partition: Detect device to be used for /data partition
Unfortunately, when an instance is booted, the NVMe devices on AWS
instances are not attached in any particular order. So what was
named "/dev/nvme1n1" during the previous boot epoch might be named
"/dev/nvme3n1" during the current boot epoch.

EBS devices do have fixed names and serial numbers, though. The
device name is determined by kdevops, but it is not exposed to the
instance. The serial number is determined by AWS, and it's reflected
in the device's symlink under /dev/disk/by-id.

The best we can do right now is use the AWS API to get the serial
number of the device named "{{ data_device }}", and then use that to
dereference the /dev/disk/by-id symlink to ensure the same EBS volume
is used for /data until a file system with a disk label lives on it.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>