create_data_partition: Detect device to be used for /data partition
When an AWS instance is booted, the NVMe devices are not attached
and recognized by the OS in any particular order. So what was named
"/dev/nvme1n1" during the previous boot epoch might be named
"/dev/nvme3n1" during the current boot epoch.

EBS devices do have fixed names and serial numbers, though. The
device name is determined by kdevops, but it is not exposed to the
instance. The serial number is determined by AWS, and it's reflected
in the device's symlink under /dev/disk/by-id on the instance.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>