terraform/AWS: Rename the Kconfig symbols for availability zones
Amazon uses the terminology "availability zone" and not
"availability region".

As we rename the Kconfig symbols, let's include the continent/nation
in these since there is, say, an "AV_ZONE_EAST_1A" in several regions.
These need to be distinct if we are to add regions and zones for other
countries.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>