terraform/AWS: Fix EBS block device naming
I noticed that when more than one instance was provisioned, the
first instance got EBS devices "/dev/sdf" and "/dev/sdg", the
second instance got devices "/dev/sdh" and "/dev/sdi", and so
on.

When provisioning multiple instances, ensure that the EBS block
device names for each instance start at /dev/sdf so that Ansible
can find the block device on which to place the data partition
(which should be /dev/sdf).

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>