terraform/AWS: Output a block_device_map
Currently the only way to access assigned device names is via an
Ansible module from the amazon.aws collection. But we can extract
the same information from our local terraform state, if we're
clever.

Subsequent patches will use this map to locate the EBS volume that
is to contain the /data file system, since a device_name is assigned
when the volume attachment is provisioned and does not change.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>