terraform/GCE: Add a VPC
I noticed that when creating more than one instance, none of them
were reachable from my Ansible controller. Time to bite the bullet
and add a network that is common to the instances. For the moment,
this will be an IPv4-only network.

Followed https://cloud.google.com/vpc/docs/vpc and some other
helpful bits of guidance.

A provider version bump is needed to get support for the
google_compute_network_attachment resource.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>