ansible.cfg: Explicitly set the ssh user
I've been seeing tasks that are marked "become: true" fail with
this message:

  Timeout (12s): waiting for privilege escalation promt:

This appears to be a common failure mode, but it is due to a broad
variety of causes. I think I've nailed this one with this addition
to ansible.cfg. It changes "ESTABLISHING SSH SESSION FOR: NONE"
to "... FOR: <user>" where <user> is always a user that is allowed
to use sudo.

The current Terraform Kconfig menu sets this user, but guestfs does
not -- the Ansible ssh user for guestfs is always "kdevops". The
ansible.cfg.j2 file reflects this.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>