ltp: Replace /opt with a symlink to data partition
The size of some cloud OS images is too small for the ltp
installation (eg, the RHEL 8 OS image on Azure is about 10GB). This
causes the ltp workflow to fail 100% of the time during the test
build step.

A simple solution is to replace /opt with a symlink to /data/opt, as
kdevops can make /data large enough to hold the install.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>